refactor: split shell widgets, move scaffold, and train dialog parts (#3878)
Extract library parts for main-menu constants, move-units dialog scaffold,
commodity-cost train widgets, game tab bar state, and CtActionTextButton
state; colocate ShellEntryAutoCenter with map-state shell helpers.

Refs #3878
