+++ app/lib/features/game/widgets/train/train_commodity_cost_dialog_base_resource_bar.dart
@@ -1,0 +1,111 @@
+part of 'train_commodity_cost_dialog_base.dart';
+
+/// Treasury / peasants / commodity chip resource bar shared by the military and
+/// naval train dialogs (icon chips, not the civilian label/value entry bar).
+class _CommodityCostResourceBar extends StatelessWidget {
+  const _CommodityCostResourceBar({
+    required this.treasury,
+    required this.remainingTreasury,
+    required this.peasants,
+    required this.remainingPeasants,
+    required this.stockpile,
+    required this.committedCommodities,
+    required this.commodityIds,
+    required this.deficitHint,
+    required this.l10n,
+  });
+
+  final int treasury;
+  final int remainingTreasury;
+  final int peasants;
+  final int remainingPeasants;
+  final Map<String, int> committedCommodities;
+  final Stockpile stockpile;
+  final List<String> commodityIds;
+  final String? deficitHint;
+  final AppLocalizations l10n;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        TrainDialogResourceBarBox(
+          child: Wrap(
+            spacing: 10,
+            runSpacing: 6,
+            children: [
+              _buildTreasuryChip(),
+              _buildPeasantsChip(),
+              for (final commodityId in commodityIds)
+                _buildCommodityChip(commodityId),
+            ],
+          ),
+        ),
+        ..._buildDeficitHint(context),
+      ],
+    );
+  }
+
+  Widget _buildTreasuryChip() {
+    return TrainDialogResourceChip(
+      child: Text(
+        l10n.trainUnits_treasury(
+          '${formatTreasuryCurrency(remainingTreasury)} / '
+          '${formatTreasuryCurrency(treasury)}',
+        ),
+      ),
+    );
+  }
+
+  Widget _buildPeasantsChip() {
+    return TrainDialogResourceChip(
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const WorkerIcon(workerType: 'peasant', size: 14),
+          const SizedBox(width: 4),
+          Flexible(
+            child: Text(
+              l10n.trainUnits_peasantsValue('$remainingPeasants / $peasants'),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildCommodityChip(String commodityId) {
+    return TrainDialogResourceChip(
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          ResourceIcon(commodityId: commodityId, size: 14),
+          const SizedBox(width: 4),
+          Flexible(
+            child: Text(
+              l10n.trainMilitary_commodityValue(
+                commodityDisplayName(commodityId),
+                '${stockpile.quantityOf(commodityId) - (committedCommodities[commodityId] ?? 0)}'
+                ' / ${stockpile.quantityOf(commodityId)}',
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  List<Widget> _buildDeficitHint(BuildContext context) {
+    if (deficitHint == null) return const [];
+    return [
+      const SizedBox(height: 4),
+      Text(
+        deficitHint!,
+        style: Theme.of(context).textTheme.bodySmall?.copyWith(
+          color: EditorialMonoclePalette.danger,
+        ),
+      ),
+    ];
+  }
+}
